fix warning in camera and fix fps display in cgal_qt5 so we don't have to overwrite it in the viewer Begin the implementation of the graphcal hints

+++ GraphicsView/src/CGAL_Qt5/camera.cpp
@@ -1664,11 +1664,11 @@ void MultiplyMatrices4by4OpenGL_GLdouble(GLdouble *result, GLdouble *matrix1, GL
       r3[2] = MAT(m, 3, 2), r3[3] = MAT(m, 3, 3),
       r3[7] = 1.0, r3[4] = r3[5] = r3[6] = 0.0;
    /* choose pivot - or die */
-   if (fabsf(r3[0]) > fabsf(r2[0]))
+   if (fabs(r3[0]) > fabs(r2[0]))
       SWAP_ROWS_GLdouble(r3, r2);
-   if (fabsf(r2[0]) > fabsf(r1[0]))
+   if (fabs(r2[0]) > fabs(r1[0]))
       SWAP_ROWS_GLdouble(r2, r1);
-   if (fabsf(r1[0]) > fabsf(r0[0]))
+   if (fabs(r1[0]) > fabs(r0[0]))
       SWAP_ROWS_GLdouble(r1, r0);
    if (0.0 == r0[0])
       return 0;
@@ -1713,9 +1713,9 @@ void MultiplyMatrices4by4OpenGL_GLdouble(GLdouble *result, GLdouble *matrix1, GL
       r3[7] -= m3 * s;
    }
    /* choose pivot - or die */
-   if (fabsf(r3[1]) > fabsf(r2[1]))
+   if (fabs(r3[1]) > fabs(r2[1]))
       SWAP_ROWS_GLdouble(r3, r2);
-   if (fabsf(r2[1]) > fabsf(r1[1]))
+   if (fabs(r2[1]) > fabs(r1[1]))
       SWAP_ROWS_GLdouble(r2, r1);
    if (0.0 == r1[1])
       return 0;
@@ -1747,7 +1747,7 @@ void MultiplyMatrices4by4OpenGL_GLdouble(GLdouble *result, GLdouble *matrix1, GL
       r3[7] -= m3 * s;
    }
    /* choose pivot - or die */
-   if (fabsf(r3[2]) > fabsf(r2[2]))
+   if (fabs(r3[2]) > fabs(r2[2]))
       SWAP_ROWS_GLdouble(r3, r2);
    if (0.0 == r2[2])
       return 0;
